centralize stats reporting
